# hellokratos/api/rider/v1/message_http.pb.go
// Code generated by protoc-gen-go-http. DO NOT EDIT.
// versions:
// - protoc-gen-go-http v2.9.2
// - protoc             v3.19.4
// source: api/rider/v1/message.proto

package v1

import (
	context "context"
	http "github.com/go-kratos/kratos/v2/transport/http"
	binding "github.com/go-kratos/kratos/v2/transport/http/binding"
)

// This is a compile-time assertion to ensure that this generated file
// is compatible with the kratos package it is being compiled against.
var _ = new(context.Context)
var _ = binding.EncodeURL

const _ = http.SupportPackageIsVersion1

const OperationMessageCreateGroup = "/api.rider.v1.Message/CreateGroup"
const OperationMessageGetGroupList = "/api.rider.v1.Message/GetGroupList"
const OperationMessageGetGroupMembers = "/api.rider.v1.Message/GetGroupMembers"
const OperationMessageGetGroupMessageList = "/api.rider.v1.Message/GetGroupMessageList"
const OperationMessageGetMessageList = "/api.rider.v1.Message/GetMessageList"
const OperationMessageGetUnreadMessageCount = "/api.rider.v1.Message/GetUnreadMessageCount"
const OperationMessageJoinGroup = "/api.rider.v1.Message/JoinGroup"
const OperationMessageLeaveGroup = "/api.rider.v1.Message/LeaveGroup"
const OperationMessageMarkMessageAsRead = "/api.rider.v1.Message/MarkMessageAsRead"
const OperationMessageSendGroupMessage = "/api.rider.v1.Message/SendGroupMessage"
const OperationMessageSendMessage = "/api.rider.v1.Message/SendMessage"

type MessageHTTPServer interface {
	// CreateGroup 创建群组
	CreateGroup(context.Context, *CreateGroupRequest) (*CreateGroupReply, error)
	// GetGroupList 获取群组列表
	GetGroupList(context.Context, *GetGroupListRequest) (*GetGroupListReply, error)
	// GetGroupMembers 获取群组成员列表
	GetGroupMembers(context.Context, *GetGroupMembersRequest) (*GetGroupMembersReply, error)
	// GetGroupMessageList 获取群聊消息列表
	GetGroupMessageList(context.Context, *GetGroupMessageListRequest) (*GetGroupMessageListReply, error)
	// GetMessageList 获取消息列表
	GetMessageList(context.Context, *GetMessageListRequest) (*GetMessageListReply, error)
	// GetUnreadMessageCount 获取未读消息数量
	GetUnreadMessageCount(context.Context, *GetUnreadMessageCountRequest) (*GetUnreadMessageCountReply, error)
	// JoinGroup 加入群组
	JoinGroup(context.Context, *JoinGroupRequest) (*JoinGroupReply, error)
	// LeaveGroup 退出群组
	LeaveGroup(context.Context, *LeaveGroupRequest) (*LeaveGroupReply, error)
	// MarkMessageAsRead 标记消息为已读
	MarkMessageAsRead(context.Context, *MarkMessageAsReadRequest) (*MarkMessageAsReadReply, error)
	// SendGroupMessage 发送群聊消息
	SendGroupMessage(context.Context, *SendGroupMessageRequest) (*SendGroupMessageReply, error)
	// SendMessage 发送消息
	SendMessage(context.Context, *SendMessageRequest) (*SendMessageReply, error)
}

func RegisterMessageHTTPServer(s *http.Server, srv MessageHTTPServer) {
	r := s.Route("/")
	r.POST("/rider/v1/message/send", _Message_SendMessage0_HTTP_Handler(srv))
	r.GET("/rider/v1/messages", _Message_GetMessageList0_HTTP_Handler(srv))
	r.GET("/rider/v1/messages/unread/count", _Message_GetUnreadMessageCount0_HTTP_Handler(srv))
	r.POST("/rider/v1/message/read", _Message_MarkMessageAsRead0_HTTP_Handler(srv))
	r.POST("/rider/v1/group/create", _Message_CreateGroup0_HTTP_Handler(srv))
	r.POST("/rider/v1/group/join", _Message_JoinGroup0_HTTP_Handler(srv))
	r.POST("/rider/v1/group/leave", _Message_LeaveGroup0_HTTP_Handler(srv))
	r.GET("/rider/v1/groups", _Message_GetGroupList0_HTTP_Handler(srv))
	r.GET("/rider/v1/group/members", _Message_GetGroupMembers0_HTTP_Handler(srv))
	r.POST("/rider/v1/group/message/send", _Message_SendGroupMessage0_HTTP_Handler(srv))
	r.GET("/rider/v1/group/messages", _Message_GetGroupMessageList0_HTTP_Handler(srv))
}

func _Message_SendMessage0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendMessageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageSendMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendMessage(ctx, req.(*SendMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendMessageReply)
		return ctx.Result(200, reply)
	}
}

func _Message_GetMessageList0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetMessageListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageGetMessageList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetMessageList(ctx, req.(*GetMessageListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetMessageListReply)
		return ctx.Result(200, reply)
	}
}

func _Message_GetUnreadMessageCount0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetUnreadMessageCountRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageGetUnreadMessageCount)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetUnreadMessageCount(ctx, req.(*GetUnreadMessageCountRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetUnreadMessageCountReply)
		return ctx.Result(200, reply)
	}
}

func _Message_MarkMessageAsRead0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in MarkMessageAsReadRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageMarkMessageAsRead)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.MarkMessageAsRead(ctx, req.(*MarkMessageAsReadRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*MarkMessageAsReadReply)
		return ctx.Result(200, reply)
	}
}

func _Message_CreateGroup0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in CreateGroupRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageCreateGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.CreateGroup(ctx, req.(*CreateGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*CreateGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Message_JoinGroup0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in JoinGroupRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageJoinGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.JoinGroup(ctx, req.(*JoinGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*JoinGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Message_LeaveGroup0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in LeaveGroupRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageLeaveGroup)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.LeaveGroup(ctx, req.(*LeaveGroupRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*LeaveGroupReply)
		return ctx.Result(200, reply)
	}
}

func _Message_GetGroupList0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetGroupListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageGetGroupList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGroupList(ctx, req.(*GetGroupListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetGroupListReply)
		return ctx.Result(200, reply)
	}
}

func _Message_GetGroupMembers0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetGroupMembersRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageGetGroupMembers)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGroupMembers(ctx, req.(*GetGroupMembersRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetGroupMembersReply)
		return ctx.Result(200, reply)
	}
}

func _Message_SendGroupMessage0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in SendGroupMessageRequest
		if err := ctx.Bind(&in); err != nil {
			return err
		}
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageSendGroupMessage)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.SendGroupMessage(ctx, req.(*SendGroupMessageRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*SendGroupMessageReply)
		return ctx.Result(200, reply)
	}
}

func _Message_GetGroupMessageList0_HTTP_Handler(srv MessageHTTPServer) func(ctx http.Context) error {
	return func(ctx http.Context) error {
		var in GetGroupMessageListRequest
		if err := ctx.BindQuery(&in); err != nil {
			return err
		}
		http.SetOperation(ctx, OperationMessageGetGroupMessageList)
		h := ctx.Middleware(func(ctx context.Context, req interface{}) (interface{}, error) {
			return srv.GetGroupMessageList(ctx, req.(*GetGroupMessageListRequest))
		})
		out, err := h(ctx, &in)
		if err != nil {
			return err
		}
		reply := out.(*GetGroupMessageListReply)
		return ctx.Result(200, reply)
	}
}

type MessageHTTPClient interface {
	// CreateGroup 创建群组
	CreateGroup(ctx context.Context, req *CreateGroupRequest, opts ...http.CallOption) (rsp *CreateGroupReply, err error)
	// GetGroupList 获取群组列表
	GetGroupList(ctx context.Context, req *GetGroupListRequest, opts ...http.CallOption) (rsp *GetGroupListReply, err error)
	// GetGroupMembers 获取群组成员列表
	GetGroupMembers(ctx context.Context, req *GetGroupMembersRequest, opts ...http.CallOption) (rsp *GetGroupMembersReply, err error)
	// GetGroupMessageList 获取群聊消息列表
	GetGroupMessageList(ctx context.Context, req *GetGroupMessageListRequest, opts ...http.CallOption) (rsp *GetGroupMessageListReply, err error)
	// GetMessageList 获取消息列表
	GetMessageList(ctx context.Context, req *GetMessageListRequest, opts ...http.CallOption) (rsp *GetMessageListReply, err error)
	// GetUnreadMessageCount 获取未读消息数量
	GetUnreadMessageCount(ctx context.Context, req *GetUnreadMessageCountRequest, opts ...http.CallOption) (rsp *GetUnreadMessageCountReply, err error)
	// JoinGroup 加入群组
	JoinGroup(ctx context.Context, req *JoinGroupRequest, opts ...http.CallOption) (rsp *JoinGroupReply, err error)
	// LeaveGroup 退出群组
	LeaveGroup(ctx context.Context, req *LeaveGroupRequest, opts ...http.CallOption) (rsp *LeaveGroupReply, err error)
	// MarkMessageAsRead 标记消息为已读
	MarkMessageAsRead(ctx context.Context, req *MarkMessageAsReadRequest, opts ...http.CallOption) (rsp *MarkMessageAsReadReply, err error)
	// SendGroupMessage 发送群聊消息
	SendGroupMessage(ctx context.Context, req *SendGroupMessageRequest, opts ...http.CallOption) (rsp *SendGroupMessageReply, err error)
	// SendMessage 发送消息
	SendMessage(ctx context.Context, req *SendMessageRequest, opts ...http.CallOption) (rsp *SendMessageReply, err error)
}

type MessageHTTPClientImpl struct {
	cc *http.Client
}

func NewMessageHTTPClient(client *http.Client) MessageHTTPClient {
	return &MessageHTTPClientImpl{client}
}

// CreateGroup 创建群组
func (c *MessageHTTPClientImpl) CreateGroup(ctx context.Context, in *CreateGroupRequest, opts ...http.CallOption) (*CreateGroupReply, error) {
	var out CreateGroupReply
	pattern := "/rider/v1/group/create"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageCreateGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGroupList 获取群组列表
func (c *MessageHTTPClientImpl) GetGroupList(ctx context.Context, in *GetGroupListRequest, opts ...http.CallOption) (*GetGroupListReply, error) {
	var out GetGroupListReply
	pattern := "/rider/v1/groups"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageGetGroupList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGroupMembers 获取群组成员列表
func (c *MessageHTTPClientImpl) GetGroupMembers(ctx context.Context, in *GetGroupMembersRequest, opts ...http.CallOption) (*GetGroupMembersReply, error) {
	var out GetGroupMembersReply
	pattern := "/rider/v1/group/members"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageGetGroupMembers))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetGroupMessageList 获取群聊消息列表
func (c *MessageHTTPClientImpl) GetGroupMessageList(ctx context.Context, in *GetGroupMessageListRequest, opts ...http.CallOption) (*GetGroupMessageListReply, error) {
	var out GetGroupMessageListReply
	pattern := "/rider/v1/group/messages"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageGetGroupMessageList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetMessageList 获取消息列表
func (c *MessageHTTPClientImpl) GetMessageList(ctx context.Context, in *GetMessageListRequest, opts ...http.CallOption) (*GetMessageListReply, error) {
	var out GetMessageListReply
	pattern := "/rider/v1/messages"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageGetMessageList))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// GetUnreadMessageCount 获取未读消息数量
func (c *MessageHTTPClientImpl) GetUnreadMessageCount(ctx context.Context, in *GetUnreadMessageCountRequest, opts ...http.CallOption) (*GetUnreadMessageCountReply, error) {
	var out GetUnreadMessageCountReply
	pattern := "/rider/v1/messages/unread/count"
	path := binding.EncodeURL(pattern, in, true)
	opts = append(opts, http.Operation(OperationMessageGetUnreadMessageCount))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "GET", path, nil, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// JoinGroup 加入群组
func (c *MessageHTTPClientImpl) JoinGroup(ctx context.Context, in *JoinGroupRequest, opts ...http.CallOption) (*JoinGroupReply, error) {
	var out JoinGroupReply
	pattern := "/rider/v1/group/join"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageJoinGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// LeaveGroup 退出群组
func (c *MessageHTTPClientImpl) LeaveGroup(ctx context.Context, in *LeaveGroupRequest, opts ...http.CallOption) (*LeaveGroupReply, error) {
	var out LeaveGroupReply
	pattern := "/rider/v1/group/leave"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageLeaveGroup))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// MarkMessageAsRead 标记消息为已读
func (c *MessageHTTPClientImpl) MarkMessageAsRead(ctx context.Context, in *MarkMessageAsReadRequest, opts ...http.CallOption) (*MarkMessageAsReadReply, error) {
	var out MarkMessageAsReadReply
	pattern := "/rider/v1/message/read"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageMarkMessageAsRead))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendGroupMessage 发送群聊消息
func (c *MessageHTTPClientImpl) SendGroupMessage(ctx context.Context, in *SendGroupMessageRequest, opts ...http.CallOption) (*SendGroupMessageReply, error) {
	var out SendGroupMessageReply
	pattern := "/rider/v1/group/message/send"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageSendGroupMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}

// SendMessage 发送消息
func (c *MessageHTTPClientImpl) SendMessage(ctx context.Context, in *SendMessageRequest, opts ...http.CallOption) (*SendMessageReply, error) {
	var out SendMessageReply
	pattern := "/rider/v1/message/send"
	path := binding.EncodeURL(pattern, in, false)
	opts = append(opts, http.Operation(OperationMessageSendMessage))
	opts = append(opts, http.PathTemplate(pattern))
	err := c.cc.Invoke(ctx, "POST", path, in, &out, opts...)
	if err != nil {
		return nil, err
	}
	return &out, nil
}