# hellokratos/openapi.yaml
...
openapi: 3.0.3
info:
    title: ""
    version: 0.0.1
paths:
...
    /rider/v1/message/read:
        post:
            tags:
                - Message
            description: 标记消息为已读
            operationId: Message_MarkMessageAsRead
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.MarkMessageAsReadRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.MarkMessageAsReadReply'
    /rider/v1/message/send:
        post:
            tags:
                - Message
            description: 发送消息
            operationId: Message_SendMessage
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: '#/components/schemas/api.rider.v1.SendMessageRequest'
                required: true
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.SendMessageReply'
    /rider/v1/messages:
        get:
            tags:
                - Message
            description: 获取消息列表
            operationId: Message_GetMessageList
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
                - name: limit
                  in: query
                  schema:
                    type: integer
                    format: int32
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetMessageListReply'
    /rider/v1/messages/unread/count:
        get:
            tags:
                - Message
            description: 获取未读消息数量
            operationId: Message_GetUnreadMessageCount
            parameters:
                - name: userId
                  in: query
                  schema:
                    type: string
            responses:
                "200":
                    description: OK
                    content:
                        application/json:
                            schema:
                                $ref: '#/components/schemas/api.rider.v1.GetUnreadMessageCountReply'
...
components:
    schemas:
...
        api.rider.v1.GetMessageListReply:
            type: object
            properties:
                messages:
                    type: array
                    items:
                        $ref: '#/components/schemas/api.rider.v1.MessageInfo'
                total:
                    type: integer
                    format: int32
            description: 获取消息列表响应
...
        api.rider.v1.GetUnreadMessageCountReply:
            type: object
            properties:
                count:
                    type: string
            description: 获取未读消息数量响应
...
        api.rider.v1.MarkMessageAsReadReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 标记消息为已读响应
        api.rider.v1.MarkMessageAsReadRequest:
            type: object
            properties:
                messageId:
                    type: string
            description: 标记消息为已读请求
        api.rider.v1.MessageInfo:
            type: object
            properties:
                id:
                    type: string
                fromId:
                    type: string
                toId:
                    type: string
                content:
                    type: string
                type:
                    type: integer
                    format: int32
                status:
                    type: integer
                    format: int32
                createdAt:
                    type: string
            description: 消息信息
...
        api.rider.v1.SendMessageReply:
            type: object
            properties:
                success:
                    type: boolean
                message:
                    type: string
            description: 发送消息响应
        api.rider.v1.SendMessageRequest:
            type: object
            properties:
                fromId:
                    type: string
                toId:
                    type: string
                content:
                    type: string
                type:
                    type: integer
                    format: int32
            description: 发送消息请求
...